refactor: Optimize project directory structure
