+++ deploy/core/core.cpp
