+++ deploy/core/core.hpp
